Configuring for c-code (#178)
* Fix tests for pure Python versions.
* Run the tests with ZODB on a modern Python version.

+++ .manylinux-install.sh
@@ -1,4 +1,6 @@
 #!/usr/bin/env bash
+# Generated from:
+# https://github.com/zopefoundation/meta/tree/master/config/c-code
 
 set -e -x
 
@@ -33,33 +35,47 @@ else
 fi
 
 export PURE_PYTHON=0
+# https://wiki.centos.org/yum-errors suggests this might help against yum
+# errors:
+yum clean all
+rm -rf /var/cache/yum/*
+# We need some libraries because we build wheels from scratch:
+yum -y install libffi-devel
 
 # Compile wheels
-cd /io/
-for variant in `ls -d /opt/python/cp{27,36,37,38,39}*`; do
-    PYBIN="$variant/bin"
-    "${PYBIN}/pip" install persistent wheel setuptools
-    "${PYBIN}/python" setup.py bdist_wheel
-    ls dist
-    if [ `uname -m` == 'aarch64' ]; then
-        # Running the test suite takes forever in
-        # emulation; an early run (using tox, which is also slow)
-        # took over an hour to build and then run the 2.7 tests and move on
-        # to the 3.5 tests. We still want to run tests, though!
-        # We don't want to distribute wheels for a platform that's
-        # completely untested. Consequently, we limit it to running
-        # in just one interpreter, the newest one on the list (which in principle
-        # should be the fastest), and we don't install the ZODB extra.
-        if [[ "${PYBIN}" == *"cp39"* ]]; then
-            "${PYBIN}/pip" install -e .[test]
-            "${PYBIN}/python" -c 'import BTrees.OOBTree; print(BTrees.OOBTree.BTree, BTrees.OOBTree.BTreePy)'
-            "${PYBIN}/python" -m unittest discover -s src
+for PYBIN in /opt/python/*/bin; do
+    if \
+       [[ "${PYBIN}" == *"cp27"* ]] || \
+       [[ "${PYBIN}" == *"cp35"* ]] || \
+       [[ "${PYBIN}" == *"cp36"* ]] || \
+       [[ "${PYBIN}" == *"cp37"* ]] || \
+       [[ "${PYBIN}" == *"cp38"* ]] || \
+       [[ "${PYBIN}" == *"cp39"* ]] || \
+       [[ "${PYBIN}" == *"cp310"* ]] ; then
+        "${PYBIN}/pip" install -e /io/
+        "${PYBIN}/pip" wheel /io/ -w wheelhouse/
+        if [ `uname -m` == 'aarch64' ]; then
+          # Running the test suite takes forever in
+          # emulation; an early run (using tox, which is also slow)
+          # took over an hour to build and then run the 2.7 tests and move on
+          # to the 3.5 tests. We still want to run tests, though!
+          # We don't want to distribute wheels for a platform that's
+          # completely untested. Consequently, we limit it to running
+          # in just one interpreter, the newest one on the list (which in principle
+          # should be the fastest), and we don't install the ZODB extra.
+          if [[ "${PYBIN}" == *"cp39"* ]]; then
+              cd /io/
+              "${PYBIN}/pip" install -e .[test]
+              "${PYBIN}/python" -c 'import BTrees.OOBTree; print(BTrees.OOBTree.BTree, BTrees.OOBTree.BTreePy)'
+              "${PYBIN}/python" -m unittest discover -s src
+              cd ..
+          fi
         fi
+        rm -rf /io/build /io/*.egg-info
     fi
-    rm -rf /io/build /io/*.egg-info
 done
 
 # Bundle external shared libraries into the wheels
-for whl in dist/BTrees*.whl; do
+for whl in wheelhouse/BTrees*.whl; do
     auditwheel repair "$whl" -w /io/wheelhouse/
 done
